重新抓取代理

+++ fetchIp.js
@@ -29,6 +29,7 @@ var initTime = new Date();
 var dateString = initTime.format("yyyyMMddhhmmss");
 
 var proxyIpRange = {start : 1, end : 62},
+    excludesProxys = [1, 11, 12, 48, 53, 54, 55, 57],
     excuteType = arguments[0],
     dirname = arguments[1] || dateString,
     startIndex = arguments[2] || proxyIpRange.start,
@@ -248,6 +249,12 @@ if( type == 0 || type == -1 ){
     (function(){
         var arg = arguments;
         if( proxyIndex <= endIndex ) {
+
+            if( tools.inArray(excludesProxys, proxyIndex) ) {
+                proxyIndex++;
+                arg.callee();
+                return;
+            }
 
             if( fs.existsSync(sourcePath)) {
                 fs.unlinkSync(sourcePath);
